+++ Semana 2/node/src/model/Appointments.ts
@@ -1,24 +1,23 @@
-import {startOfHour,parseISO} from 'date-fns'
-import { uuid } from "UuidV4";
+import { startOfHour, parseISO } from "date-fns";
+import { Entity, Column, PrimaryGeneratedColumn } from "typeorm";
 
 // A model é responsavel pela estrutura do dado que a aplicacao utiliza
-
-class  Appointments {
+@Entity("appointments")
+class Appointments {
+  @PrimaryGeneratedColumn("uuid")
   id: string;
+
+  @Column()
   provider: string;
+
+  @Column('timestamp with time zone')
   date: Date;
 
-  constructor({provider,date}:Omit<Appointments,"id">){
-    this.id = uuid();
-    this.provider = provider;
-    this.date = date;
-  }
-  
+  // Com o typeorm nao é necessario criar um construtor, porem para parar de dar erro é necessario desabilitar no tsconfig a propriedade strictPropertyInitialization
 }
 
- export function getParsedHour(date:string){
-  return startOfHour(parseISO(date))
+export function getParsedHour(date: string) {
+  return startOfHour(parseISO(date));
 }
 
-
-export default Appointments
+export default Appointments;